pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easyai-client"
version = "0.1.0"
description = "EasyAI user client."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "PyYAML>=6.0",
  "rich>=13.7.0",
]

[project.scripts]
easyai = "pyai_assistant.cli.client:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
